# docs/api/clario-dr-service.openapi.yaml
openapi: 3.1.0
info:
  title: ClarioDR Service API
  version: 1.0.0
  summary: Disaster recovery control-plane and agent-ingest contract.
  description: |
    Phase 1 contract foundation for the existing ClarioDR API. The gateway
    exposes the tenant control plane at `/api/v1/dr` and gates it with the
    `suite.datastream` entitlement. Mutating control-plane writes stage DR
    lifecycle events through `backend/internal/events/outbox` before commit.
  x-contract-phase: phase-1-foundation
  x-gateway-contract:
    id: clario-dr-service
    version: 1.0.0
    api_version: v1
    fail_closed: false
    fail_closed_profiles:
      standard: false
      regulated: true
    regulated_posture:
      env:
        - DR_DEPLOYMENT_PROFILE=regulated
        - DR_REGULATED_MODE=true
      startup_validation: fail_closed
      readiness_check: dr_runtime_config
      required_runtime:
        - Vault Transit
        - MinIO WORM recovery-point bucket
        - Redis leader election
        - Kafka event bus
        - mTLS agent ingest
        - failover driver
        - integration credential encryption key
        - command recovery driver
  x-route-sources:
    - backend/cmd/clario-dr-service/main.go
    - backend/internal/dr/handler/handler.go
    - backend/internal/dr/enroll/handler.go
    - backend/internal/dr/ingest/handler.go
    - backend/cmd/clario-dr-service/resilience.go
    - backend/cmd/clario-dr-service/intelligence.go
    - backend/cmd/clario-dr-service/orchestration.go
...
servers:
...
security:
  - bearerAuth: []
    tenantContext: []
...
paths:
...
  /rehearsal-proofs/gameday/{runID}:
    get:
      tags: [RehearsalProof]
      operationId: getDRGameDayRehearsalProof
      summary: Build a proof envelope from a real game-day run.
      x-required-permission: dr:read
      parameters:
        - $ref: "#/components/parameters/RunID"
      responses:
        "200":
          description: Rehearsal proof envelope.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RehearsalProofEnvelope"
        "404":
          $ref: "#/components/responses/NotFound"
  /rehearsal-proofs/runbook-runs/{runID}:
    get:
      tags: [RehearsalProof]
      operationId: getDRRunbookRehearsalProof
      summary: Build a proof envelope from a real runbook execution.
      x-required-permission: dr:read
      parameters:
        - $ref: "#/components/parameters/RunID"
      responses:
        "200":
          description: Rehearsal proof envelope.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RehearsalProofEnvelope"
        "404":
          $ref: "#/components/responses/NotFound"
  /rehearsal-proofs/failover-runs/{runID}:
    get:
      tags: [RehearsalProof]
      operationId: getDRFailoverDrillRehearsalProof
      summary: Build a proof envelope from a real failover drill run.
      x-required-permission: dr:read
      parameters:
        - $ref: "#/components/parameters/RunID"
      responses:
        "200":
          description: Rehearsal proof envelope.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RehearsalProofEnvelope"
        "404":
          $ref: "#/components/responses/NotFound"
        "422":
          $ref: "#/components/responses/UnprocessableEntity"
  /rehearsal-proofs/gameday/{runID}/sealed:
    get:
      tags: [RehearsalProof]
      operationId: getDRSealedGameDayRehearsalProof
      summary: Get the persisted, signed, WORM-sealed proof for a game-day run.
      x-required-permission: dr:read
      parameters:
        - $ref: "#/components/parameters/RunID"
      responses:
        "200":
          description: Sealed rehearsal proof and ledger inclusion reference.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DRSealedRehearsalProofEnvelope"
        "404":
          $ref: "#/components/responses/NotFound"
  /rehearsal-proofs/gameday/{runID}/seal:
    post:
      tags: [RehearsalProof]
      operationId: sealDRGameDayRehearsalProof
      summary: Assemble, sign, WORM-seal, ledger-anchor, and persist a game-day proof.
      x-required-permission: dr:write
      parameters:
        - $ref: "#/components/parameters/RunID"
      responses:
        "201":
          description: Sealed rehearsal proof.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DRSealedRehearsalProofEnvelope"
        "400":
          $ref: "#/components/responses/BadRequest"
        "404":
          $ref: "#/components/responses/NotFound"
        "409":
          $ref: "#/components/responses/Conflict"
        "422":
          $ref: "#/components/responses/UnprocessableEntity"
  /rehearsal-proofs/runbook-runs/{runID}/sealed:
    get:
      tags: [RehearsalProof]
      operationId: getDRSealedRunbookRehearsalProof
      summary: Get the persisted, signed, WORM-sealed proof for a runbook execution.
      x-required-permission: dr:read
      parameters:
        - $ref: "#/components/parameters/RunID"
      responses:
        "200":
          description: Sealed rehearsal proof and ledger inclusion reference.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DRSealedRehearsalProofEnvelope"
        "404":
          $ref: "#/components/responses/NotFound"
  /rehearsal-proofs/runbook-runs/{runID}/seal:
    post:
      tags: [RehearsalProof]
      operationId: sealDRRunbookRehearsalProof
      summary: Assemble, sign, WORM-seal, ledger-anchor, and persist a runbook proof.
      x-required-permission: dr:write
      parameters:
        - $ref: "#/components/parameters/RunID"
      responses:
        "201":
          description: Sealed rehearsal proof.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DRSealedRehearsalProofEnvelope"
        "400":
          $ref: "#/components/responses/BadRequest"
        "404":
          $ref: "#/components/responses/NotFound"
        "409":
          $ref: "#/components/responses/Conflict"
        "422":
          $ref: "#/components/responses/UnprocessableEntity"
  /rehearsal-proofs/failover-runs/{runID}/sealed:
    get:
      tags: [RehearsalProof]
      operationId: getDRSealedFailoverRehearsalProof
      summary: Get the persisted, signed, WORM-sealed proof for a failover drill.
      x-required-permission: dr:read
      parameters:
        - $ref: "#/components/parameters/RunID"
      responses:
        "200":
          description: Sealed rehearsal proof and ledger inclusion reference.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DRSealedRehearsalProofEnvelope"
        "404":
          $ref: "#/components/responses/NotFound"
  /rehearsal-proofs/failover-runs/{runID}/seal:
    post:
      tags: [RehearsalProof]
      operationId: sealDRFailoverRehearsalProof
      summary: Assemble, sign, WORM-seal, ledger-anchor, and persist a failover drill proof.
      x-required-permission: dr:write
      parameters:
        - $ref: "#/components/parameters/RunID"
      responses:
        "201":
          description: Sealed rehearsal proof.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DRSealedRehearsalProofEnvelope"
        "400":
          $ref: "#/components/responses/BadRequest"
        "404":
          $ref: "#/components/responses/NotFound"
        "409":
          $ref: "#/components/responses/Conflict"
        "422":
          $ref: "#/components/responses/UnprocessableEntity"
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
    tenantContext:
      type: apiKey
      in: header
      name: X-Tenant-ID
      description: Optional fallback when tenant context is not present in the JWT.
    agentMutualTLS:
      type: mutualTLS
  parameters:
...
    RunID:
      name: runID
      in: path
      required: true
      schema:
        type: string
        format: uuid
...
  responses:
    BadRequest:
      description: Invalid request.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
...
    NotFound:
      description: Resource not found.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
...
    Conflict:
      description: Invalid state or already existing resource.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    UnprocessableEntity:
      description: Request was refused by a safety or semantic guard.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
...
  schemas:
    ErrorResponse:
      type: object
      required: [code, message]
      properties:
        status:
          type: integer
        code:
          type: string
        message:
          type: string
        details: true
        request_id:
          type: string
    UUID:
      type: string
      format: uuid
    Timestamp:
      type: string
      format: date-time
...
    DRSealedRehearsalProof:
      type: object
      required: [id, tenant_id, subject_kind, subject_run_id, envelope, envelope_hash, signature, signature_alg, signed_at, created_at]
      properties:
        id:
          $ref: "#/components/schemas/UUID"
        tenant_id:
          $ref: "#/components/schemas/UUID"
        subject_kind:
          type: string
        subject_run_id:
          type: string
        envelope:
          $ref: "#/components/schemas/RehearsalProof"
        envelope_hash:
          type: string
          description: SHA-256 digest of the canonical proof envelope.
        signature:
          type: string
          description: Detached digital signature over the envelope hash.
        signature_alg:
          type: string
        signed_at:
          $ref: "#/components/schemas/Timestamp"
        worm_object_key:
          type: string
        ledger_seq:
          type: integer
          format: int64
        ledger_entry_hash:
          type: string
        created_at:
          $ref: "#/components/schemas/Timestamp"
    DRSealedRehearsalProofEnvelope:
      type: object
      required: [data]
      properties:
        data:
          $ref: "#/components/schemas/DRSealedRehearsalProof"
    RehearsalRunRef:
      type: object
      required: [type, id]
      properties:
        type:
          type: string
          enum: [gameday, runbook_studio, failover_drill, runbook_run]
        id:
          $ref: "#/components/schemas/UUID"
        group_id:
          $ref: "#/components/schemas/UUID"
        scenario_id:
          $ref: "#/components/schemas/UUID"
        runbook_id:
          $ref: "#/components/schemas/UUID"
        mode:
          type: string
        status:
          type: string
    RehearsalScope:
      type: object
      required: [name]
      properties:
        name:
          type: string
        environment:
          type: string
        system_ids:
          type: array
          items:
            type: string
    RehearsalSystemRef:
      type: object
      required: [id]
      properties:
        id:
          type: string
        name:
          type: string
        kind:
          type: string
        critical:
          type: boolean
    RehearsalRunbookRef:
      type: object
      properties:
        id:
          $ref: "#/components/schemas/UUID"
        name:
          type: string
        version:
          type: string
        content_hash:
          type: string
        source:
          type: string
    RehearsalTargets:
      type: object
      properties:
        rto_target_seconds:
          type: integer
        rto_actual_seconds:
          type: integer
        rpo_target_seconds:
          type: integer
        rpo_actual_seconds:
          type: integer
        rto_met:
          type: boolean
        rpo_met:
          type: boolean
    RehearsalApprovalRef:
      type: object
      properties:
        id:
          type: string
        action:
          type: string
        approver:
          type: string
        approved_at:
          $ref: "#/components/schemas/Timestamp"
        source:
          type: string
    RehearsalHealthCheck:
      type: object
      required: [name, status, passed, checked_at]
      properties:
        name:
          type: string
        target:
          type: string
        status:
          type: string
          enum: [passed, failed, warning, unknown, running, skipped, complete]
        passed:
          type: boolean
        checked_at:
          $ref: "#/components/schemas/Timestamp"
        detail:
          type: string
        evidence_id:
          type: string
        finished_at:
          $ref: "#/components/schemas/Timestamp"
    RehearsalIntegrityVerdict:
      type: object
      required: [source, verdict, passed, checked_at]
      properties:
        source:
          type: string
        verdict:
          type: string
          enum: [passed, failed, warning, unknown]
        passed:
          type: boolean
        checked_at:
          $ref: "#/components/schemas/Timestamp"
        evidence_id:
          type: string
        detail:
          type: string
    RehearsalTeardownStatus:
      type: object
      required: [status]
      properties:
        status:
          type: string
          enum: [complete, incomplete, not_started, unknown]
        failback_id:
          type: string
        completed_at:
          $ref: "#/components/schemas/Timestamp"
        detail:
          type: string
    RehearsalEvidenceReportRef:
      type: object
      required: [id, hash, algorithm]
      properties:
        id:
          type: string
        hash:
          type: string
        algorithm:
          type: string
        generated_at:
          $ref: "#/components/schemas/Timestamp"
        source:
          type: string
    RehearsalProvenance:
      type: object
      required: [source, live, seeded, demo, captured_at]
      properties:
        source:
          type: string
          enum: [live_rehearsal, seeded_demo, imported, unknown]
        live:
          type: boolean
        seeded:
          type: boolean
        demo:
          type: boolean
        captured_from:
          type: string
        recorded_by:
          type: string
        captured_at:
          $ref: "#/components/schemas/Timestamp"
        notes:
          type: string
    RehearsalProof:
      type: object
      required:
        - schema_version
        - id
        - tenant_id
        - run
        - scope
        - systems
        - runbook
        - targets
        - approval_refs
        - started_at
        - health_checks
        - integrity_verdicts
        - failback_teardown
        - overall_verdict
        - provenance
        - envelope_hash
        - generated_at
      properties:
        schema_version:
          type: string
          const: clario.dr.rehearsal_proof.v1
        id:
          $ref: "#/components/schemas/UUID"
        tenant_id:
          $ref: "#/components/schemas/UUID"
        run:
          $ref: "#/components/schemas/RehearsalRunRef"
        scope:
          $ref: "#/components/schemas/RehearsalScope"
        systems:
          type: array
          items:
            $ref: "#/components/schemas/RehearsalSystemRef"
        runbook:
          $ref: "#/components/schemas/RehearsalRunbookRef"
        targets:
          $ref: "#/components/schemas/RehearsalTargets"
        approval_refs:
          type: array
          items:
            $ref: "#/components/schemas/RehearsalApprovalRef"
        started_at:
          $ref: "#/components/schemas/Timestamp"
        completed_at:
          $ref: "#/components/schemas/Timestamp"
        health_checks:
          type: array
          items:
            $ref: "#/components/schemas/RehearsalHealthCheck"
        integrity_verdicts:
          type: array
          items:
            $ref: "#/components/schemas/RehearsalIntegrityVerdict"
        failback_teardown:
          $ref: "#/components/schemas/RehearsalTeardownStatus"
        evidence_report:
          $ref: "#/components/schemas/RehearsalEvidenceReportRef"
        overall_verdict:
          type: string
          enum: [passed, failed, warning, unknown]
        provenance:
          $ref: "#/components/schemas/RehearsalProvenance"
        envelope_hash:
          type: string
        generated_at:
          $ref: "#/components/schemas/Timestamp"
...
    RehearsalProofEnvelope:
      $ref: "#/components/schemas/DataEnvelopeRehearsalProof"
...
    DataEnvelopeRehearsalProof:
      type: object
      required: [data]
      properties:
        data:
          $ref: "#/components/schemas/RehearsalProof"
...